requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
prometheus-fastapi-instrumentator
boto3
